mcpdemo/pyproject.toml
[project]
name = "mcpdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.13",
    "aiosignal>=1.4.0",
    "annotated-types>=0.7.0",
    "anyio>=4.9.0",
    "attrs>=25.3.0",
    "backoff>=2.2.1",
    "bcrypt>=4.3.0",
    "cachetools>=5.5.2",
    "certifi>=2025.6.15",
    "charset-normalizer>=3.4.2",
    "click>=8.2.1",
    "colorama>=0.4.6",
    "distro>=1.9.0",
    "fastapi>=0.116.0",
    "fastapi-mcp>=0.3.4",
    "frozenlist>=1.7.0",
    "google-api-core>=2.25.1",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "googleapis-common-protos>=1.70.0",
    "greenlet>=3.2.3",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httplib2>=0.22.0",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.1",
    "idna>=3.10",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "jsonschema>=4.24.0",
    "jsonschema-pydantic>=0.6",
    "jsonschema-specifications>=2025.4.1",
    "langchain>=0.3.26",
    "langchain-core>=0.3.68",
    "langchain-groq>=0.3.5",
    "langchain-openai>=0.3.27",
    "langchain-text-splitters>=0.3.8",
    "langsmith>=0.4.4",
    "markdown-it-py>=3.0.0",
    "mcp-use>=1.3.3",
    "mcp[cli]>=1.10.1",
    "mdurl>=0.1.2",
    "multidict>=6.6.3",
    "oauthlib>=3.3.1",
    "openai>=1.93.0",
    "orjson>=3.10.18",
    "packaging>=24.2",
    "passlib>=1.7.4",
    "pip>=25.1.1",
    "posthog>=6.0.2",
    "propcache>=0.3.2",
    "proto-plus>=1.26.1",
    "protobuf>=6.31.1",
    "pyasn1>=0.6.1",
    "pyasn1-modules>=0.4.2",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pydantic-settings>=2.10.1",
    "pygments>=2.19.2",
    "pyparsing>=3.2.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "referencing>=0.36.2",
    "requests>=2.32.4",
    "requests-oauthlib>=2.0.0",
    "requests-toolbelt>=1.0.0",
    "rich>=14.0.0",
    "rpds-py>=0.26.0",
    "rsa>=4.9.1",
    "scarf-sdk>=0.1.2",
    "shellingham>=1.5.4",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.41",
    "sse-starlette>=2.3.6",
    "starlette>=0.46.2",
    "tenacity>=9.1.2",
    "tomli>=2.2.1",
    "typer>=0.16.0",
    "typing-extensions>=4.14.0",
    "typing-inspection>=0.4.1",
    "uritemplate>=4.2.0",
    "urllib3>=2.5.0",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
    "yarl>=1.20.1",
    "zstandard>=0.23.0",
]
